+++ src/components/radio/radio.ts
@@ -32,7 +32,23 @@ export const radioPropTypes = {
   fullWidth: {
     type: Boolean,
     default: false,
-  }
+  },
+  displayHelper: {
+    type: Boolean,
+    default: false,
+  },
+  helperIcon: {
+    type: String,
+    default: null,
+  },
+  helperText: {
+    type: String,
+    default: '',
+  },
+  error: {
+    type: Boolean,
+    default: false,
+  },
 };
 
 export const radioEmitTypes = ['update:modelValue'];