+++ docs/locales/en/guide/changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.22.3 (2025-11-19)
+
+- Fix:
+  - Added Display Helper support for Radio component.
+  - Update Documentation for Radio Component.
+    ([#35957](https://dev.azure.com/sproutphil/Sprout%20Design%20System/_git/Sprout%20Design%20System%20Next/pullrequest/36031) by @jarias)
+
 ## 2.22.2 (2025-11-19)
 
 - Fix: